Simplify custom geometry textures.
Don't allow custom offsets for textured cube face textures -- just
start at 0,0 and use the section of the texture needed to fill the face
based on the dimensions of the cube. This means we can't pack multiple
textures into one 16x16 image, but png compression makes that not matter
much.